fix(extractor): Correct a cache key bug and prevent root_names reuse (#46)
