+++ .changeset/pink-pens-tickle.md
@@ -1,0 +1,5 @@
+---
+"@svelte-docgen/extractor": patch
+---
+
+Fixed a cache key bug and resolved an issue caused by root_names reuse
