+++ apps/docs/package.json
@@ -16,7 +16,7 @@
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-cloudflare": "^4.9.0",
-		"@sveltejs/kit": "^2.13.0",
+		"@sveltejs/kit": "^2.14.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@sveltepress/theme-default": "^5.0.5",
 		"@sveltepress/vite": "^1.1.2",
